Defined BookingListView. Bookings sorted and then displayed in template. Createad template and applied some bootsrap.
