+++ trainWellApp/templates/staff/booking_list.html
@@ -1,0 +1,74 @@
+{% extends isajax|yesno:"ajax_base.html,base.html" %}
+{% load app_filters %}
+{% load static %}
+
+{% block body_block %}
+
+    <section id="day_schedule" class="container-fluid">
+
+        <div id="header" class="row px-5 pt-5 pb-3">
+            <div class="col">
+                <div class="btn-group" role="group">
+                    <button id="pw" type="button" class="btn btn-secondary btn-sm d-inline"
+                            onclick="swipeDay('{% url 'staff:booking_list' %}','{{ prev_day }}')">
+                        <i class="fas fa-chevron-left"> </i>
+                    </button>
+                    <button id="nw" type="button" class="btn btn-secondary btn-sm d-inline"
+                            onclick="swipeDay('{% url 'staff:booking_list' %}','{{ next_day }}')">
+                        <i class="fas fa-chevron-right"> </i>
+                    </button>
+                </div>
+                <button id="today" type="button" class="btn btn-secondary btn-sm mr-3"
+                        onclick="swipeDay('{% url 'staff:booking_list' %}','{% now "d/m/Y" %}')">Today
+                </button>
+                <button id="today" type="button" class="btn btn-success btn-sm mr-3"
+                        onclick="">
+                    <i class="fas fa-plus mr-2"></i>Incidence
+                </button>
+            </div>
+            <h5>{{ day|date:"D, d F Y" }}</h5>
+        </div>
+
+        {% for h in hours %}
+            {% now "H:00" as curr_hour %}
+            {% if curr_hour == h %}
+                <div style="border-top-style: solid; border-top-width: 1.5px; border-top-color: #e74a3b;"
+                     class="row m-5 animated slideInRight">
+            {% else %}
+                <div class="row m-5 animated slideInRight border-top">
+            {% endif %}
+                <small class="col-sm-12 mr-3 text-muted">{{ h }}</small>
+
+                {% for key, vals in  object_list.items %}
+
+                    {% if h == vals.values|tolist|first|first %}
+                        <div class="card col-auto ml-5 mt-3 mb-2 border-0 shadow-sm"
+                             style="border-left: .5rem solid #cc73e1 !important">
+                            <div class="card-body">
+                                <h6 class="card-title">{{ key.name }}</h6>
+                                <h6 class="card-subtitle mb-2 text-muted">{{ key.event }}</h6>
+                                <div class="card-text btn-group" role="group">
+                                    {% for k, v in vals.items %}
+                                        <button type="button" class="btn btn-sm btn-light" data-toggle="popover"
+                                        title="Hours" data-trigger="hover" data-content="{{ v }}"
+                                                style="font-size: small">{{ k }}</button>
+                                    {% endfor %}
+                                </div>
+                            </div>
+                        </div>
+                    {% endif %}
+                {% endfor %}
+            </div>
+        {% endfor %}
+
+        <script id="popover_script" type="text/javascript">
+            $(document).ready(function () {
+                $('[data-toggle="popover"]').popover();
+            });
+        </script>
+
+    </section>
+
+{% endblock %}
+
+
