+++ trainWellApp/urls/staff_urls.py
@@ -6,5 +6,4 @@ app_name = "trainWellApp"
 
 urlpatterns = [
     path('bookings/', BookingListView.as_view(), name='booking_list'),
-
 ]
